+++ audioswitch/src/main/java/com/twilio/audioswitch/bluetooth/BluetoothHeadsetConnectionListener.kt
@@ -1,5 +1,6 @@
 package com.twilio.audioswitch.bluetooth
 
 internal interface BluetoothHeadsetConnectionListener {
-    fun onBluetoothHeadsetStateChanged()
+    fun onBluetoothHeadsetStateChanged(headsetName: String? = null)
+    fun onBluetoothHeadsetActivationError()
 }