+++ audioswitch/src/main/java/com/twilio/audioswitch/selection/AudioDeviceSelector.kt
@@ -3,14 +3,10 @@ package com.twilio.audioswitch.selection
 import android.bluetooth.BluetoothAdapter
 import android.content.Context
 import android.media.AudioManager
-import com.twilio.audioswitch.android.BluetoothIntentProcessorImpl
 import com.twilio.audioswitch.android.BuildWrapper
 import com.twilio.audioswitch.android.LogWrapper
-import com.twilio.audioswitch.bluetooth.BluetoothController
-import com.twilio.audioswitch.bluetooth.BluetoothHeadsetCacheManager
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetConnectionListener
 import com.twilio.audioswitch.bluetooth.BluetoothHeadsetManager
-import com.twilio.audioswitch.bluetooth.BluetoothHeadsetReceiver
 import com.twilio.audioswitch.selection.AudioDevice.BluetoothHeadset
 import com.twilio.audioswitch.selection.AudioDevice.Earpiece
 import com.twilio.audioswitch.selection.AudioDevice.Speakerphone
@@ -47,53 +43,44 @@ class AudioDeviceSelector {
                         AudioFocusRequestWrapper())
         this.logger = logger
         this.audioDeviceManager = audioDeviceManager
-        val deviceCache = BluetoothHeadsetCacheManager(logger)
-        this.headsetCache = deviceCache
         this.wiredHeadsetReceiver = WiredHeadsetReceiver(context, logger)
-        this.bluetoothController = BluetoothAdapter.getDefaultAdapter()?.let { bluetoothAdapter ->
-            BluetoothController(context,
-                    bluetoothAdapter,
-                    BluetoothHeadsetManager(logger, bluetoothAdapter, deviceCache),
-                    BluetoothHeadsetReceiver(context, logger, BluetoothIntentProcessorImpl(),
-                            audioDeviceManager, deviceCache)
-            )
-        } ?: run {
-            logger.d(TAG, "Bluetooth is not supported on this device")
-            null
-        }
+        this.bluetoothHeadsetManager = BluetoothHeadsetManager.newInstance(context, logger,
+                BluetoothAdapter.getDefaultAdapter(), audioDeviceManager)
     }
 
     internal constructor(
         logger: LogWrapper,
         audioDeviceManager: AudioDeviceManager,
         wiredHeadsetReceiver: WiredHeadsetReceiver,
-        bluetoothController: BluetoothController?,
-        headsetCache: BluetoothHeadsetCacheManager
+        headsetManager: BluetoothHeadsetManager?
     ) {
         this.logger = logger
         this.audioDeviceManager = audioDeviceManager
         this.wiredHeadsetReceiver = wiredHeadsetReceiver
-        this.bluetoothController = bluetoothController
-        this.headsetCache = headsetCache
+        this.bluetoothHeadsetManager = headsetManager
     }
 
     private var logger: LogWrapper = LogWrapper()
-    private var headsetCache: BluetoothHeadsetCacheManager? = null
     private val audioDeviceManager: AudioDeviceManager
     private val wiredHeadsetReceiver: WiredHeadsetReceiver
-    internal val bluetoothController: BluetoothController?
     internal var audioDeviceChangeListener: AudioDeviceChangeListener? = null
     private var selectedDevice: AudioDevice? = null
     private var userSelectedDevice: AudioDevice? = null
     private var wiredHeadsetAvailable = false
     private val mutableAudioDevices = ArrayList<AudioDevice>()
+    private var bluetoothHeadsetManager: BluetoothHeadsetManager? = null
 
     internal var state: State = STOPPED
     internal enum class State {
         STARTED, ACTIVATED, STOPPED
     }
     internal val bluetoothDeviceConnectionListener = object : BluetoothHeadsetConnectionListener {
-        override fun onBluetoothHeadsetStateChanged() {
+        override fun onBluetoothHeadsetStateChanged(headsetName: String?) {
+            enumerateDevices(headsetName)
+        }
+
+        override fun onBluetoothHeadsetActivationError() {
+            if (userSelectedDevice is BluetoothHeadset) userSelectedDevice = null
             enumerateDevices()
         }
     }
@@ -125,7 +112,7 @@ class AudioDeviceSelector {
         audioDeviceChangeListener = listener
         when (state) {
             STOPPED -> {
-                bluetoothController?.start(bluetoothDeviceConnectionListener)
+                bluetoothHeadsetManager?.start(bluetoothDeviceConnectionListener)
                 wiredHeadsetReceiver.start(wiredDeviceConnectionListener)
                 enumerateDevices()
                 state = STARTED
@@ -182,15 +169,15 @@ class AudioDeviceSelector {
         when (audioDevice) {
             is BluetoothHeadset -> {
                 audioDeviceManager.enableSpeakerphone(false)
-                bluetoothController?.activate()
+                bluetoothHeadsetManager?.activate()
             }
             is Earpiece, is WiredHeadset -> {
                 audioDeviceManager.enableSpeakerphone(false)
-                bluetoothController?.deactivate()
+                bluetoothHeadsetManager?.deactivate()
             }
             is Speakerphone -> {
                 audioDeviceManager.enableSpeakerphone(true)
-                bluetoothController?.deactivate()
+                bluetoothHeadsetManager?.deactivate()
             }
         }
     }
@@ -202,7 +189,7 @@ class AudioDeviceSelector {
     fun deactivate() {
         when (state) {
             ACTIVATED -> {
-                bluetoothController?.deactivate()
+                bluetoothHeadsetManager?.deactivate()
 
                 // Restore stored audio state
                 audioDeviceManager.restoreAudioState()
@@ -241,10 +228,16 @@ class AudioDeviceSelector {
      */
     val availableAudioDevices: List<AudioDevice> = mutableAudioDevices
 
-    private fun enumerateDevices() {
+    private fun enumerateDevices(bluetoothHeadsetName: String? = null) {
         mutableAudioDevices.clear()
-        headsetCache?.cachedHeadsets?.let { devices ->
-            if (devices.isNotEmpty()) mutableAudioDevices.add(devices.last())
+        /*
+         * Since the there is a delay between receiving the ACTION_ACL_CONNECTED event and receiving
+         * the name of the connected device from querying the BluetoothHeadset proxy class, the
+         * headset name received from the ACTION_ACL_CONNECTED intent needs to be passed into this
+         * function.
+         */
+        bluetoothHeadsetManager?.getHeadset(bluetoothHeadsetName)?.let {
+            mutableAudioDevices.add(it)
         }
         if (wiredHeadsetAvailable) {
             mutableAudioDevices.add(WiredHeadset())
@@ -267,7 +260,17 @@ class AudioDeviceSelector {
         selectedDevice = if (userSelectedDevice != null && userSelectedDevicePresent(mutableAudioDevices)) {
             userSelectedDevice
         } else if (mutableAudioDevices.size > 0) {
-            mutableAudioDevices[0]
+            val firstAudioDevice = mutableAudioDevices[0]
+            /*
+             * If there was an error starting bluetooth sco, then the selected AudioDevice should
+             * be the next valid device in the list.
+             */
+            if (firstAudioDevice is BluetoothHeadset &&
+                    bluetoothHeadsetManager?.hasActivationError() == true) {
+                mutableAudioDevices[1]
+            } else {
+                firstAudioDevice
+            }
         } else {
             null
         }
@@ -297,7 +300,7 @@ class AudioDeviceSelector {
     }
 
     private fun closeListeners() {
-        bluetoothController?.stop()
+        bluetoothHeadsetManager?.stop()
         wiredHeadsetReceiver.stop()
         audioDeviceChangeListener = null
         state = STOPPED