+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-### 0.1.6
+### 0.2.0
+
+Enhancements
+- Added support for multiple connected bluetooth headsets.
+  - The library will now accurately display the up to date active bluetooth headset within the `AudiodDeviceSelector` `availableAudioDevices` and `selectedAudioDevice` functions.
+    - Other connected headsets are not stored by the library at this moment.
+  - In the event of a failure to connecting audio to a bluetooth headset, the library will revert the selected audio device (this is usually the Earpiece on a phone).
+  - If a user would like to switch between multiple Bluetooth headsets, then they need to switch the active bluetooth headset from the system Bluetooth settings.
+    - The newly activated headset will be propagated to the `AudiodDeviceSelector` `availableAudioDevices` and `selectedAudioDevice` functions.
 
 Bug Fixes